alterando tipagem peso de int para float

+++ include/Grafo.h
@@ -40,8 +40,8 @@ public:
     virtual int get_num_vizinhos(int id) { return 0; };                             // Retorna o numero de vizinhos de um vertice
     virtual void dfs(int v, bool* visitado){};                                      // Realiza a busca em profundidade
 
-    virtual void adicionar_vertice(int id, int peso = 0){};                         // Adiciona um vertice no grafo
-    virtual void adicionar_aresta(int origem, int destino, int peso = 1){};         // Adiciona uma aresta no grafo
+    virtual void adicionar_vertice(int id, float peso = 0.0){};                         // Adiciona um vertice no grafo
+    virtual void adicionar_aresta(int origem, int destino, float peso = 1.0){};         // Adiciona uma aresta no grafo
     virtual void remover_primeira_aresta(int id){};                                 // Remove a primeira aresta de um vertice
     virtual void remover_vertice(int id){};                                         // Remove um vertice do grafo
     virtual void remover_aresta(int origem, int destino){};                         // Remove uma aresta do grafo
